Avatar for now

+++ src/main/resources/liquibase/scripts/user.sql
@@ -22,7 +22,7 @@ CREATE TABLE comment
 );
 CREATE TABLE users
 (
-    id BIGINT PRIMARY KEY,
+    id INT PRIMARY KEY,
     username TEXT,
     first_name TEXT,
     last_name  TEXT,
